Added Authentications in Front End and Backend & Authorization in Backend

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post("", checkAuth,multer({storage: storage}).single("image"), (req, res,
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save()   // save() is provided by mongoose package itself
   .then(createdPost => {
@@ -97,14 +98,24 @@ router.put("/:id", checkAuth,multer({storage: storage}).single("image"), (req,re
   }
   const post = new Post({ _id: req.body.Id, title: req.body.title, content: req.body.content, imagePath: imagePath })
   console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: "Update Successfully"});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    console.log(result)
+    if(result.nModified > 0) {
+      res.status(200).json({message: "Update Successfully"});
+    } else {
+      res.status(401).json({message: "Not Authorised"});
+    }
   })
 });
 
 router.delete("/:id",checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    res.status(200).json({message: "Post Deleted"});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    console.log('deleted', result);
+    if(result.deletedCount > 1) {
+      res.status(200).json({message: "Post Deleted"});
+    } else {
+      res.status(401).json({message: "Not Authorised"});
+    }
   });
 });
